+++ src/client/options/disconnect.rs
@@ -9,5 +9,33 @@ pub struct Options {
 
     /// The session expiry interval property. Not allowed to be set to a non-zero value
     /// if the session expiry interval property in the CONNECT packet has been 0.
+    /// This value overrides the session expiry interval negotiated in the handshake.
     pub session_expiry_interval: Option<SessionExpiryInterval>,
+}
+
+impl Default for Options {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Options {
+    /// Creates new disconnect options with will publication disabled and no session expiry interval.
+    pub const fn new() -> Self {
+        Self {
+            publish_will: false,
+            session_expiry_interval: None,
+        }
+    }
+
+    /// Sets the publish will flag to true.
+    pub const fn publish_will(mut self) -> Self {
+        self.publish_will = true;
+        self
+    }
+    /// Sets the session expiry interval property.
+    pub const fn session_expiry_interval(mut self, interval: SessionExpiryInterval) -> Self {
+        self.session_expiry_interval = Some(interval);
+        self
+    }
 }