+++ src/client/options/mod.rs
@@ -4,8 +4,10 @@ mod connect;
 mod disconnect;
 mod publish;
 mod subscribe;
+mod will;
 
-pub use connect::{Options as ConnectOptions, WillOptions};
+pub use connect::Options as ConnectOptions;
 pub use disconnect::Options as DisconnectOptions;
 pub use publish::{Options as PublicationOptions, TopicReference};
 pub use subscribe::{Options as SubscriptionOptions, RetainHandling};
+pub use will::Options as WillOptions;