+++ src/v5/packet/connect.rs
@@ -120,6 +120,7 @@ impl<'p> ConnectPacket<'p> {
         keep_alive: KeepAlive,
         session_expiry_interval: SessionExpiryInterval,
         receive_maximum: u16,
+        request_response_information: bool,
     ) -> Self {
         Self {
             will_retain: false,
@@ -130,7 +131,9 @@ impl<'p> ConnectPacket<'p> {
             receive_maximum: ReceiveMaximum(receive_maximum),
             maximum_packet_size: MaximumPacketSize::default(),
             topic_alias_maximum: None,
-            request_response_information: None,
+            request_response_information: request_response_information
+                .then_some(true)
+                .map(Into::into),
             request_problem_information: None,
             authentication_method: None,
             authentication_data: None,
@@ -188,12 +191,11 @@ mod unit {
     use crate::{
         config::{KeepAlive, SessionExpiryInterval},
         test::tx::encode,
-        types::{MqttBinary, MqttString, QoS, Will},
+        types::{MqttBinary, MqttString, QoS, TopicName, Will},
         v5::{
             packet::ConnectPacket,
             property::{
-                ContentType, CorrelationData, MessageExpiryInterval, PayloadFormatIndicator,
-                ResponseTopic, WillDelayInterval,
+                ContentType, MessageExpiryInterval, PayloadFormatIndicator, WillDelayInterval,
             },
         },
     };
@@ -207,6 +209,7 @@ mod unit {
             KeepAlive::Seconds(7439),
             SessionExpiryInterval::EndOnDisconnect,
             u16::MAX,
+            false,
         );
 
         #[rustfmt::skip]
@@ -239,6 +242,7 @@ mod unit {
             KeepAlive::Seconds(6789),
             SessionExpiryInterval::EndOnDisconnect,
             u16::MAX,
+            false,
         );
 
         packet.add_user_name(MqttString::try_from("Franz").unwrap());
@@ -299,23 +303,20 @@ mod unit {
             KeepAlive::Seconds(6789),
             SessionExpiryInterval::Seconds(893475),
             u16::MAX,
+            true,
         );
 
         packet.add_user_name(MqttString::try_from("Franz").unwrap());
         packet.add_password(MqttBinary::try_from("security!".as_bytes()).unwrap());
         packet.add_will(
             Will {
-                will_topic: MqttString::try_from("dead").unwrap(),
+                will_topic: TopicName::new_checked(MqttString::try_from("dead").unwrap()).unwrap(),
                 will_delay_interval: Some(WillDelayInterval(234589)),
                 payload_format_indicator: Some(PayloadFormatIndicator(true)),
                 message_expiry_interval: Some(MessageExpiryInterval(84807612)),
                 content_type: Some(ContentType(MqttString::try_from("text/plain").unwrap())),
-                response_topic: Some(ResponseTopic(
-                    MqttString::try_from("calling/undertaker").unwrap(),
-                )),
-                correlation_data: Some(CorrelationData(
-                    MqttBinary::try_from([68, 98, 204, 101].as_slice()).unwrap(),
-                )),
+                response_topic: None,
+                correlation_data: None,
                 will_payload: MqttBinary::try_from([12, 8, 98].as_slice()).unwrap(),
             },
             QoS::ExactlyOnce,
@@ -326,7 +327,7 @@ mod unit {
         encode!(packet,
             [
                 0x10,       // Packet type
-                0x68,       // Remaining length
+                0x4E,       // Remaining length
                 0x00,       // Protocol name len MSB
                 0x04,       // Protocol name len LSB
                 b'M',       // Protocol name
@@ -338,9 +339,11 @@ mod unit {
                 0x1A,       // Keep alive MSB
                 0x85,       // Keep alive LSB
 
-                0x05,       // Property length
+                0x07,       // Property length
 
                 0x11, 0x00, 0x0D, 0xA2, 0x23, // Session expiry interval
+
+                0x19, 0x01, // Request Response Information
 
                 0x00,       // Client identifier len MSB
                 0x03,       // Client identifier len LSB
@@ -348,7 +351,7 @@ mod unit {
                 b'b',       //
                 b'a',       // Client identifier
 
-                0x35,       // Will properties length
+                0x19,       // Will properties length
 
                 0x18, 0x00, 0x03, 0x94, 0x5D, // Will delay interval
 
@@ -358,13 +361,6 @@ mod unit {
 
                 0x03, 0x00, 0x0A, // Content type
                 b't', b'e', b'x', b't', b'/', b'p', b'l', b'a', b'i', b'n', // Content type
-
-                0x08,       // Response topic
-                0x00, 0x12,
-                b'c', b'a', b'l', b'l', b'i', b'n', b'g', b'/', 
-                b'u', b'n', b'd', b'e', b'r', b't', b'a', b'k', b'e', b'r', // Response topic
-
-                0x09, 0x00, 0x04, 68, 98, 204, 101, // Correlation data value
 
                 0x00,       // Will topic len MSB
                 0x04,       // Will topic len LSB