+++ tests/integration/mod.rs
@@ -2,5 +2,6 @@ mod creds;
 mod message_retry;
 mod pub_properties;
 mod pub_sub;
+mod request_response;
 mod session;
 mod sub_options;